updates assets

+++ Unit.java
@@ -29,77 +29,62 @@ public abstract class Unit {
     }
 
     public int getCost() {
-
         return cost;
     }
 
     public void setCost(int cost) {
-
         this.cost = cost;
     }
 
     public int getMin_power() {
-
         return min_power;
     }
 
     public void setMin_power(int min_power) {
-
         this.min_power = min_power;
     }
 
     public int getMax_power() {
-
         return max_power;
     }
 
     public void setMax_power(int max_power) {
-
         this.max_power = max_power;
     }
 
-    public double getPriorityATT() {
-
+    public int getPriorityATT() {
         return priorityATT;
     }
 
     public void setPriorityATT(int priorityATT) {
-
         this.priorityATT = priorityATT;
     }
 
-    public double getPriorityDEF() {
-
+    public int getPriorityDEF() {
         return priorityDEF;
     }
 
     public void setPriorityDEF(int priorityDEF) {
-
         this.priorityDEF = priorityDEF;
     }
 
     public int getMpt() {
-
         return mpt;
     }
 
     public void setMpt(int mpt) {
-
         this.mpt = mpt;
     }
 
     public int getCpt() {
-
         return cpt;
     }
 
     public void setCpt(int cpt) {
-
         this.cpt = cpt;
     }
 
     public void addCpt() {
-
         this.cpt++;
     }
 }
